pyproject.toml
[tool.poetry]
name = "torrent"
version = "0.1.0"
description = ""
authors = ["azat715 <azat715@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"bencoder.pyx" = "^2.0.1"
fire = "^0.3.1"
peewee = "^3.14.0"

[tool.poetry.scripts]
torrent = "torrent.main:cli"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-datafiles = "^2.0"
pylint = "^2.6.0"
black = "^20.8b1"
mypy = "^0.790"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
